+++ src/server/Main.hs
@@ -1,35 +1,28 @@
-{-# LANGUAGE OverloadedStrings #-}
-
 module Main where
 
-import Control.Concurrent (threadDelay)
-import Control.Monad (forever)
-import Data.ByteString qualified as BS
-import Data.ByteString.Lazy qualified as LB
-import Data.Text qualified as Text
-import Data.Text.Encoding qualified as Text
-import System.IO
+import Control.Monad.Reader (runReaderT)
+import Data.Aeson qualified as Json
+
+import Recalc.Server
+import Recalc.Server.Protocol
+
+type EngineState = ()
+
+type SheetsApi = ToApi SpreadsheetProtocol
 
 main :: IO ()
-main = do
-  mapM_ (`hSetBuffering` NoBuffering) [stdin, stdout]
-  hSetBuffering stderr LineBuffering
-
-  forever $ do
-    sendBS "\"Hello, from Haskell!\""
-    threadDelay 1000_000
-    hPutStrLn stderr "Hello, from Haskell!"
-    threadDelay 1000_000
-
--- | send json-rpc message on stdout
-sendBS :: LB.ByteString -> IO ()
-sendBS = BS.putStr . BS.toStrict . toRpc
+main = runHandler @SheetsApi () $ \state ->
+  hoist @SheetsApi (`runReaderT` state) (namedHandlers server)
  where
-  toRpc bs =
-    LB.concat
-      [ "Content-Length: " <> showLB (LB.length bs)
-      , "\r\n\r\n"
-      , bs
-      ]
+  server :: SpreadsheetProtocol (AsServerT (Handler EngineState))
+  server =
+    SpreadsheetProtocol
+      { rpcOpen = Main.rpcOpen
+      , rpcClose = Main.rpcClose
+      }
 
-  showLB = LB.fromStrict . Text.encodeUtf8 . Text.pack . show
+rpcOpen :: (Maybe Id, OpenParams) -> Handler EngineState Json.Value
+rpcOpen _ = fail "'open' not implemented"
+
+rpcClose :: (Maybe Id, CloseParams) -> Handler EngineState ()
+rpcClose _ = fail "'close' not implemented"