+++ recalc-vscode/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Revision history for recalc-vscode
 
-## 0.0.2 -- 2025-02-06
+## 0.0.3 -- 2025-02-07
 
 * Initial version: create project structure for a custom editor.
 * Add generic rpc implementation, client implementation and logging facilities.
+* Integrate type-safe json-rpc communication between client and server.